Initial rsync build script and workflow
Using my android-dropbear build script and workflow as a starting point,
create a build script for cross-compiling rsync for Android on Linux and
a GitHub action to run the build script after each push.

The workflow also creates a release with the built binaries if one
doesn't already exist for the latest release.