Manually disable x86_64-specific ASM optmisations
The configure logic in rsync 3.2.0 for deciding if x86_64 optimised
assembly should be built checks the type of the CPU being used to build
rather than the target CPU.  As we are generally building on x86_64 but
targeting ARM or ARM64 processors the check passes but the build fails.

Pass the --disable-simd flag to configure to ensure this is disabled.
It looks like this flag might not be necessary in rsync 3.2.1 as
configure on master now uses a compile test which should hopefully cause
the ASM to be disabled automatically.