Fix triggering of build from version update PRs
The branches-ignore entry for update-version added in bce0ccb558 fixed
the issue of both pull_request and push triggering separate workflow
runs when a version update PR has additional commits pushed.  However,
it also stopped the version update PRs triggering the build workflow at
all when created as actions by the actions user don't trigger further
workflow runs (but the push used to trigger a workflow as it runs as my
user via a deploy key).

Attempt to fix this regression by removing the branch from
branches-ignore and adding a condition to the Compile job to suppress
the push event when a user who isn't actions pushes to the
update-version branch.