Switch from GitHub Packages to Maven Central
GitHub Packages Maven registry doesn't allow anonymous (e.g.
unauthenticated) access and GitHub no-longer appears to be planning to
add support for it.  This means that consuming packages is annoying due
to always having to use credentials or tokens for package downloads.

To resolve this, switch publishing the package to Maven Central instead.
This is a bit more involved as they require the JAR and POM to be
supplied with PGP signatures as well as requiring more metadata to be
defined in the POM.  To make the implementation cleaner, create a new
pom.xml file with the required metadata and use Maven to build, sign and
publish the JAR.

Also configure Dependabot to update the Maven plugins and add a link to
the package at Maven Central in the README.

Resolves #42.